+++ tests/unit_tests/steps/test_grab.py
@@ -7,7 +7,8 @@ from pathlib import Path
 from types import SimpleNamespace
 from unittest import mock
 
-from fab.steps.grab import GrabFolder, GrabFcm
+from fab.steps.grab.fcm import GrabFcm
+from fab.steps.grab.folder import GrabFolder
 
 
 class TestGrabFolder(object):
@@ -42,7 +43,7 @@ class TestGrabFcm(object):
 
         mock_config = SimpleNamespace(source_root=source_root)
         with mock.patch('pathlib.Path.mkdir'):
-            with mock.patch('fab.steps.grab.run_command') as mock_run:
+            with mock.patch('fab.steps.grab.fcm.run_command') as mock_run:
                 grabber.run(artefact_store={}, config=mock_config)
 
         mock_run.assert_called_once_with(['fcm', 'export', '--force', source_url, str(source_root / dst_label)])
@@ -56,7 +57,7 @@ class TestGrabFcm(object):
 
         mock_config = SimpleNamespace(source_root=source_root)
         with mock.patch('pathlib.Path.mkdir'):
-            with mock.patch('fab.steps.grab.run_command') as mock_run:
+            with mock.patch('fab.steps.grab.fcm.run_command') as mock_run:
                 grabber.run(artefact_store={}, config=mock_config)
 
         mock_run.assert_called_once_with(
